Move localFile class definition into new header file drive_local.h
I was running into circular header dependencies for what I wanted to do next.
