+++ src/dos/drive_local.cpp
@@ -23,6 +23,7 @@
 // #define DEBUG 1
 
 #include "drives.h"
+#include "drive_local.h"
 
 #include <cerrno>
 #include <climits>
